refactor: move progress Language into runner
